refactor: split libvirt backend & workflow-contract modules into API-stable packages (#50)
