+++ implementations/python/packages/raes_backend_libvirt/techvault_native/_driver.py
@@ -17,26 +17,22 @@ from urllib.parse import urlsplit
 
 from raes_contracts.diagnostics import Diagnostic
 
-from ._techvault_native_helpers import default_connector as _default_connector
-from ._techvault_native_helpers import default_kernel_path as _default_kernel_path
-from ._techvault_native_ops import (
+from .._techvault_native_helpers import default_connector as _default_connector
+from .._techvault_native_helpers import default_kernel_path as _default_kernel_path
+from .._techvault_native_ops import (
     _CODE_OPERATION_FAILED,
     _CODE_OWNERSHIP_CONFLICT,
-    _CODE_READBACK_FAILED,
     _CODE_RESIDUAL_STATE,
     _CODE_UNAVAILABLE,
     _DEFAULT_CONNECTION_URI,
     _artifact_token,
-    _call,
     _diagnostic,
-    _ensure_name_available,
-    _NativeResource,
 )
 
 if TYPE_CHECKING:
     from raes_contracts.realization_envelope import BackendRealizationEnvelopeModel
 
-from .driver import (
+from ..driver import (
     DomainHandle,
     DomainSpec,
     DriverResult,
@@ -44,66 +40,37 @@ from .driver import (
     NetworkSpec,
     RealizationObservation,
 )
-from .drivers.libvirt import Connector, _existing_uuid, _raes_uuid
-from .envelopes import load_libvirt_realization_envelope
-from .techvault_appliance import (
-    BusyboxInitramfsBuilder,
-    InitramfsBuilder,
-    copy_kernel_for_libvirt,
-    make_libvirt_readable,
-)
-from .techvault_concerns import techvault_observation_diagnostics, techvault_spec_diagnostics
-from .techvault_lifecycle import (
+from ..drivers.libvirt import Connector, _existing_uuid, _raes_uuid
+from ..envelopes import load_libvirt_realization_envelope
+from ..techvault_appliance import BusyboxInitramfsBuilder, InitramfsBuilder
+from ..techvault_concerns import techvault_observation_diagnostics, techvault_spec_diagnostics
+from ..techvault_lifecycle import (
     NativeOwnershipConflict as _OwnershipConflict,
 )
-from .techvault_lifecycle import (
+from ..techvault_lifecycle import (
     deactivate_and_undefine as _deactivate_and_undefine,
 )
-from .techvault_lifecycle import (
+from ..techvault_lifecycle import (
     resolve_native as _resolve_native,
 )
-from .techvault_lifecycle import (
+from ..techvault_lifecycle import (
     verify_native_removed as _verify_native_removed,
 )
-from .techvault_matrix import (
-    as_sequence as _as_sequence,
-)
-from .techvault_matrix import (
+from ..techvault_matrix import (
     domain_xml as _domain_xml,
 )
-from .techvault_matrix import (
+from ..techvault_matrix import (
     native_matrix as _native_matrix,
 )
-from .techvault_matrix import (
-    network_xml as _network_xml,
-)
-from .techvault_matrix import (
+from ..techvault_matrix import (
     safe_name as _safe_name,
 )
-from .techvault_observation import (
+from ..techvault_observation import (
     canonical_digest,
-    domain_observations,
     file_digest,
-    network_observations,
     snapshot_from_observations,
 )
-from .techvault_probe import (
-    NativeLibvirtProbe,
-    ProbeResult,
-    check_native_readiness,
-    expected_surface,
-    native_soc_readback,
-)
-
-__all__ = [
-    "BusyboxInitramfsBuilder",
-    "NativeLibvirtProbe",
-    "ProbeResult",
-    "TechVaultNativeLibvirtDriver",
-    "check_native_readiness",
-    "expected_surface",
-    "native_soc_readback",
-]
+from ._define import define_domain, define_domains, define_network, define_networks
 
 
 @dataclass
@@ -285,81 +252,17 @@ class TechVaultNativeLibvirtDriver:
     def _define_networks(
         self, connection: object, matrix: Mapping[str, object]
     ) -> tuple[list[NetworkHandle], list[Diagnostic], list[RealizationObservation]]:
-        handles: list[NetworkHandle] = []
-        diagnostics: list[Diagnostic] = []
-        observations: list[RealizationObservation] = []
-        for network in _as_sequence(matrix.get("networks")):
-            if isinstance(network, Mapping):
-                handle, diagnostic, observed = self._define_network(connection, network)
-                if handle is not None:
-                    handles.append(handle)
-                if diagnostic is not None:
-                    diagnostics.append(diagnostic)
-                observations.extend(observed)
-                if diagnostic is not None:
-                    break
-        return handles, diagnostics, observations
+        return define_networks(self, connection, matrix)
 
     def _define_network(
         self, connection: object, network: Mapping[str, object]
     ) -> tuple[NetworkHandle | None, Diagnostic | None, tuple[RealizationObservation, ...]]:
-        address = str(network.get("address", ""))
-        native: _NativeResource | None = None
-        handle: NetworkHandle | None = None
-        diagnostic: Diagnostic | None = None
-        observations: tuple[RealizationObservation, ...] = ()
-        self._names[address] = str(network.get("runtime_name", ""))
-        try:
-            _ensure_name_available(
-                connection,
-                "networkLookupByName",
-                str(network.get("runtime_name", "")),
-                address,
-            )
-            native = _call(connection, "networkDefineXML", _network_xml(network))
-            if not self.define_only:
-                native.create()
-        except _OwnershipConflict:
-            self._names.pop(address, None)
-            diagnostic = _diagnostic(_CODE_OWNERSHIP_CONFLICT, address)
-        except Exception:
-            if native is None:
-                self._names.pop(address, None)
-            else:
-                self._realized.add(address)
-                handle = NetworkHandle(address=address)
-            diagnostic = _diagnostic(_CODE_OPERATION_FAILED, address)
-        else:
-            self._realized.add(address)
-            handle = NetworkHandle(address=address, realized=True)
-            try:
-                observations = network_observations(native, network)
-            except Exception:
-                diagnostic = _diagnostic(_CODE_READBACK_FAILED, address)
-        return handle, diagnostic, observations
+        return define_network(self, connection, network)
 
     def _define_domains(
         self, connection: object, matrix: Mapping[str, object]
     ) -> tuple[list[DomainHandle], list[Diagnostic], list[RealizationObservation]]:
-        handles: list[DomainHandle] = []
-        diagnostics: list[Diagnostic] = []
-        observations: list[RealizationObservation] = []
-        network_addresses = {
-            str(item.get("runtime_name", "")): str(item.get("address", ""))
-            for item in _as_sequence(matrix.get("networks"))
-            if isinstance(item, Mapping)
-        }
-        for domain in _as_sequence(matrix.get("domains")):
-            if isinstance(domain, Mapping):
-                handle, diagnostic, observed = self._define_domain(connection, domain, network_addresses)
-                if handle is not None:
-                    handles.append(handle)
-                if diagnostic is not None:
-                    diagnostics.append(diagnostic)
-                observations.extend(observed)
-                if diagnostic is not None:
-                    break
-        return handles, diagnostics, observations
+        return define_domains(self, connection, matrix)
 
     def _define_domain(
         self,
@@ -367,57 +270,7 @@ class TechVaultNativeLibvirtDriver:
         domain: Mapping[str, object],
         network_addresses: Mapping[str, str],
     ) -> tuple[DomainHandle | None, Diagnostic | None, tuple[RealizationObservation, ...]]:
-        address = str(domain.get("address", ""))
-        native: _NativeResource | None = None
-        handle: DomainHandle | None = None
-        diagnostic: Diagnostic | None = None
-        observations: tuple[RealizationObservation, ...] = ()
-        self._names[address] = str(domain.get("runtime_name", ""))
-        try:
-            _ensure_name_available(
-                connection,
-                "lookupByName",
-                str(domain.get("runtime_name", "")),
-                address,
-            )
-            kernel = copy_kernel_for_libvirt(
-                self.kernel_path,
-                self.state_dir / "kernel" / f"{_artifact_token(address)}-{self.kernel_path.name}",
-            )
-            initrd = self.initramfs_builder.build(
-                domain=domain,
-                target=self.state_dir / "initramfs" / f"{_artifact_token(address)}.cpio.gz",
-            )
-            make_libvirt_readable(initrd)
-            self._artifacts[address] = (kernel, initrd)
-            native = _call(connection, "defineXML", self._render_domain_xml(domain, kernel=kernel, initrd=initrd))
-            if not self.define_only:
-                native.create()
-        except _OwnershipConflict:
-            self._names.pop(address, None)
-            diagnostic = _diagnostic(_CODE_OWNERSHIP_CONFLICT, address)
-        except Exception:
-            if native is None:
-                self._cleanup_artifacts(address)
-                self._names.pop(address, None)
-            else:
-                self._realized.add(address)
-                handle = DomainHandle(address=address)
-            diagnostic = _diagnostic(_CODE_OPERATION_FAILED, address)
-        else:
-            self._realized.add(address)
-            handle = DomainHandle(address=address, realized=True)
-            try:
-                observations = domain_observations(
-                    native,
-                    domain,
-                    network_addresses,
-                    kernel=kernel,
-                    initrd=initrd,
-                )
-            except Exception:
-                diagnostic = _diagnostic(_CODE_READBACK_FAILED, address)
-        return handle, diagnostic, observations
+        return define_domain(self, connection, domain, network_addresses)
 
     def destroy(
         self,
